rough draft for dag project endpoints
