+++ Makefile
@@ -50,7 +50,7 @@ build-daemons:
 	docker compose -f docker/docker-compose.yml build daemons
 
 build-docs:
-	$(POETRY) mkdocs build --config-file config/mkdocs.yml
+	$(UV) mkdocs build --config-file config/mkdocs.yml
 	docker compose -f docker/docker-compose.yml build docs
 
 build-main:
@@ -98,7 +98,7 @@ run-postgres:
 		echo "Waiting for postgres..."; \
 		sleep 1; \
 	done
-	$(POETRY) alembic upgrade head
+	$(UV) alembic upgrade head
 # ----------------------- AWS Commands -----------------------
 # TODO: Use hyphens instead of underscores
 # https://www.gnu.org/software/libc/manual/html_node/Argument-Syntax.html
@@ -146,18 +146,18 @@ local-docs:
 	mkdocs serve --config-file config/mkdocs.yml
 
 local-api:
-	$(POETRY) fastapi dev webapp/api/server.py --port 8001
+	$(UV) fastapi dev webapp/api/server.py --port 8001
 
 local-main:
-	$(POETRY) fastapi dev webapp/main/server.py
+	$(UV) fastapi dev webapp/main/server.py
 
 local-auth:
-	$(POETRY) fastapi dev webapp/auth/server.py --port 8002
+	$(UV) fastapi dev webapp/auth/server.py --port 8002
 
 # Note that for webhook functionality, you will need to use a service like ngrok to expose your local server to the internet. 
 # I run `ngrok http 8000`, and then use the forwarding URL as the webhook URL in the GitHub app settings. See webapp/daemons/README.md for more details.
 local-daemons:
-	/bin/bash -c "set -a; source .env.daemons; set +a; cd webapp/daemons && $(POETRY) fastapi dev server.py"
+	/bin/bash -c "set -a; source .env.daemons; set +a; cd webapp/daemons && $(UV) fastapi dev server.py"
 
 
 # Build Packages
