+++ src/concrete-core/concrete/orchestrators/__init__.py
@@ -1,8 +1,4 @@
+from .base import Orchestrator
 from .software_orchestrator import SoftwareOrchestrator
-
-
-class Orchestrator:
-    pass
-
 
 __all__ = ["SoftwareOrchestrator", "Orchestrator"]
