+++ webapp/api/pyproject.toml
@@ -6,7 +6,7 @@ readme = "README.md"
 requires-python = ">=3.11.9, <4.0.0"
 dependencies = [
     "concrete-db",
-    "fastapi>=0.115.4",
+    "fastapi[standard]>=0.115.4",
 ]
 
 [tool.uv.sources]
